+++ modules/alloy/src/test/ts/browser/behaviour/DockingTest.ts
@@ -1,4 +1,4 @@
-import { ApproxStructure, Assertions, Cleaner, Logger, Step, Waiter } from '@ephox/agar';
+import { ApproxStructure, Assertions, Cleaner, Logger, Step, Waiter, GeneralSteps } from '@ephox/agar';
 import { UnitTest } from '@ephox/bedrock-client';
 import { window } from '@ephox/dom-globals';
 import { DomEvent, Element } from '@ephox/sugar';
@@ -33,9 +33,6 @@ UnitTest.asynctest('DockingTest', (success, failure) => {
             },
             containerBehaviours: Behaviour.derive([
               Docking.config({
-                leftAttr: 'data-dock-left',
-                topAttr: 'data-dock-top',
-                positionAttr: 'data-dock-pos',
                 onDocked: store.adder('static.onDocked'),
                 onUndocked: store.adder('static.onUndocked')
               })
@@ -49,14 +46,11 @@ UnitTest.asynctest('DockingTest', (success, failure) => {
                 background: 'red',
                 position: 'absolute',
                 top: '2300px',
-                left: '200px'
+                right: '200px'
               }
             },
             containerBehaviours: Behaviour.derive([
               Docking.config({
-                leftAttr: 'data-dock-left',
-                topAttr: 'data-dock-top',
-                positionAttr: 'data-dock-pos',
                 onDocked: store.adder('absolute.onDocked'),
                 onUndocked: store.adder('absolute.onUndocked')
               })
@@ -78,7 +72,11 @@ UnitTest.asynctest('DockingTest', (success, failure) => {
       return ApproxStructure.build((s, str, arr) => {
         return s.element('div', {
           styles: {
-            position: str.none()
+            position: str.none(),
+            left: str.none(),
+            top: str.none(),
+            right: str.none(),
+            bottom: str.none()
           }
         });
       });
@@ -94,19 +92,32 @@ UnitTest.asynctest('DockingTest', (success, failure) => {
       });
     };
 
-    return [
-      // On initial load, it should have no position.
+    const sAssertInitialStructure = Logger.t('Assert initial structure', GeneralSteps.sequence([
       Assertions.sAssertStructure(
-        'On initial load, static box should have neither position: absolute nor position: fixed',
+        'Assert initial structure of staticBox. Box should have neither "position: absolute" nor "position: fixed"',
         boxWithNoPosition(),
         staticBox.element()
       ),
       Assertions.sAssertStructure(
-        'On initial load, absolute box should have position: absolute',
-        boxWithPosition('absolute'),
+        'Assert initial structure of absoluteBox',
+        ApproxStructure.build((s, str, arr) => {
+          return s.element('div', {
+            styles: {
+              position: str.is('absolute'),
+              left: str.none(),
+              top: str.is('2300px'),
+              right: str.is('200px'),
+              bottom: str.none(),
+            }
+          });
+        }),
         absoluteBox.element()
-      ),
-      store.sAssertEq('Before docking', [ ]),
+      )
+    ]));
+
+    return [
+      store.sAssertEq('Store should be empty', [ ]),
+      sAssertInitialStructure,
 
       Logger.t(
         'Scroll completely offscreen',
@@ -130,6 +141,25 @@ UnitTest.asynctest('DockingTest', (success, failure) => {
           boxWithPosition('fixed'),
           absoluteBox.element()
         )
+      ),
+      // For future reference - Docking is always using 'left' and 'top' when docked but this behavior isn't set in stone
+      Logger.t(
+        'When fixed, absoluteBox should be positioned with "top" and "left"',
+        Assertions.sAssertStructure(
+          'Assert structure of absoluteBox',
+          ApproxStructure.build((s, str, arr) => {
+            return s.element('div', {
+              styles: {
+                position: str.is('fixed'),
+                left: str.contains('px'), // assert isSome
+                top: str.contains('0px'),
+                right: str.none(),
+                bottom: str.none()
+              }
+            });
+          }),
+          absoluteBox.element()
+        ),
       ),
       store.sAssertEq('When docked', [ 'static.onDocked', 'absolute.onDocked' ]),
       store.sClear,
@@ -158,6 +188,10 @@ UnitTest.asynctest('DockingTest', (success, failure) => {
         )
       ),
       store.sAssertEq('After undocked', [ 'static.onUndocked', 'absolute.onUndocked' ]),
+      Logger.t(
+        'After undocking, the structure of the docked elements should be what it originally was',
+        sAssertInitialStructure
+      )
     ];
   }, cleanup.wrap(success), cleanup.wrap(failure));
 });